Updates (UI + Manager + ScoreSystem)
I have added a small update in the scoring system so that it is easy to select how many points should be awarded per obstacle dodged.
I have created a GameManager that holds important information about the game (for now just the time passed). I have created a timer and finished the important feedback UI elements for the player

Also I have changed something in the PlayerHealth script which prevents anyone from setting the health of the player to any value below 1.